+++ src/Benchmarks/MicroBench.Current/SerializeBenchmarks.cs
@@ -24,73 +24,73 @@ namespace Microbench
     public class SerializeBenchmarks
     {
         [Benchmark]
-        public int Serialize_StringTable_SingleString()
+        public long Serialize_StringTable_SingleString()
         {
             return StringTable.Serializer.Write(Constants.Buffers.StringTable_WithString, Constants.StringTables.WithString);
         }
 
         [Benchmark]
-        public int Serialize_StringTable_Vector()
+        public long Serialize_StringTable_Vector()
         {
             return StringTable.Serializer.Write(Constants.Buffers.StringTable_WithVector, Constants.StringTables.WithVector);
         }
 
         [Benchmark]
-        public int Serialize_StringTable_EmptyTable()
+        public long Serialize_StringTable_EmptyTable()
         {
             return StringTable.Serializer.Write(Constants.Buffers.Stringtable_Empty, Constants.StringTables.Empty);
         }
 
         [Benchmark]
-        public int Serialize_PrimitivesTable_Empty()
+        public long Serialize_PrimitivesTable_Empty()
         {
             return PrimitivesTable.Serializer.Write(Constants.Buffers.PrimitivesTable_Empty, Constants.PrimitiveTables.Empty);
         }
 
         [Benchmark]
-        public int Serialize_PrimitivesTable_Full()
+        public long Serialize_PrimitivesTable_Full()
         {
             return PrimitivesTable.Serializer.Write(Constants.Buffers.PrimitivesTable_Full, Constants.PrimitiveTables.Full);
         }
 
         [Benchmark]
-        public int Serialize_StructTable_SingleRef()
+        public long Serialize_StructTable_SingleRef()
         {
             return StructsTable.Serializer.Write(Constants.Buffers.StructTable_SingleRef, Constants.StructTables.SingleRef);
         }
 
         [Benchmark]
-        public int Serialize_StructTable_SingleValue()
+        public long Serialize_StructTable_SingleValue()
         {
             return StructsTable.Serializer.Write(Constants.Buffers.StructTable_SingleValue, Constants.StructTables.SingleValue);
         }
 
         [Benchmark]
-        public int Serialize_StructTable_VecRef()
+        public long Serialize_StructTable_VecRef()
         {
             return StructsTable.Serializer.Write(Constants.Buffers.StructTable_VecRef, Constants.StructTables.VectorRef);
         }
 
         [Benchmark]
-        public int Serialize_StructTable_VecValue()
+        public long Serialize_StructTable_VecValue()
         {
             return StructsTable.Serializer.Write(Constants.Buffers.StructTable_VecValue, Constants.StructTables.VectorValue);
         }
 
         [Benchmark]
-        public int Serialize_SafeUnion()
+        public long Serialize_SafeUnion()
         {
             return UnionTable.Serializer.Write(Constants.Buffers.UnionTable_Safe, Constants.UnionTables.Safe);
         }
 
         [Benchmark]
-        public int Serialize_UnsafeUnion()
+        public long Serialize_UnsafeUnion()
         {
             return UnionTable.Serializer.Write(Constants.Buffers.UnionTable_Unsafe, Constants.UnionTables.Unsafe);
         }
 
         [Benchmark]
-        public int Serialize_MixedUnion()
+        public long Serialize_MixedUnion()
         {
             return UnionTable.Serializer.Write(Constants.Buffers.UnionTable_Mixed, Constants.UnionTables.Mixed);
         }
