+++ src/Tests/FlatSharpEndToEndTests/Oracle/AlignmentTests.cs
@@ -75,9 +75,9 @@ public class AlignmentTests
         };
 
         Span<byte> memory = new byte[10240];
-        int offset = FS.AlignmentTestOuterHoder.Serializer.Write(memory, holder);
+        long offset = FS.AlignmentTestOuterHoder.Serializer.Write(memory, holder);
 
-        var parsed = Flatc.AlignmentTestOuterHoder.GetRootAsAlignmentTestOuterHoder(new ByteBuffer(memory.Slice(0, offset).ToArray()));
+        var parsed = Flatc.AlignmentTestOuterHoder.GetRootAsAlignmentTestOuterHoder(new ByteBuffer(memory.Slice(0, (int)offset).ToArray()));
 
         var outer = parsed.Value.Value;
         Assert.AreEqual(1, outer.A);
